requirements.txt
annotated-types
audioread
dtwalign
eng_to_ipa
epitran 
faster-whisper
flask
flask_cors
omegaconf
openai-whisper
ortools
pandas
pickle-mixin
requests
sacremoses # suggested by marian translation model
sentencepiece
silero
soundfile 
sqlalchemy
structlog
-f https://download.pytorch.org/whl/torch_stable.html
torch 
torchaudio 
transformers
